Reset ruby-version to latest installed patch on developer machine
Ruby core team is working on issues relating to OpenSSL updates and OS X
Mavericks, but they have yet to release patches. The current latest
patches do not play well with Apple's post-Heartbleed OpenSSL update.
For now, we'll use whatever patch-level is highest on the box running
the app, whether it's development or production.