Remove dotenv gem from development and test environments
Dotenv will load settings from the `.env` file in development. This
file should be used to save settings for heroku deploys. In the future,
we should try setting things up to leverage a different `.env` file
locally, for example `.env.development`. The `.env` file should be
reserved for pushing settings changes to heroku.